Version 0.2.0:
Refactored core log handling architecture.
- Simplified LogFile to handle only physical file access.
- Introduced RecordSource abstraction for logical record access.
- Added CSVRecordSource and CSVRecordParser for CSV support.
- Organized includes using IWYU.

+++ libhawk/src/platform/file_mapping_posix.cpp
@@ -6,8 +6,9 @@
 #include <sys/mman.h>
 #include <sys/stat.h>
 #include <unistd.h>
-
+#include <cstddef>
 #include <stdexcept>
+#include <string>
 
 namespace hawk::platform {
 
@@ -20,7 +21,7 @@ void FileMapping::open(const std::string& path) {
     if (fstat(fd_, &st) != 0)
         throw std::runtime_error("Failed to stat file");
 
-    size_ = st.st_size;
+    size_ = static_cast<std::size_t>(st.st_size);
 
     data_ = static_cast<const char*>(
         mmap(nullptr, size_, PROT_READ, MAP_PRIVATE, fd_, 0));
